# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "LINUX DO Credit",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/admin/users": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "username",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/users/{id}/status": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "状态",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.updateUserStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "user.updateUserStatusRequest": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                }
            }
        },
...
        "util.ResponseAny": {
            "type": "object",
            "properties": {
                "data": {},
                "error_msg": {
                    "type": "string",
                    "example": ""
                }
            }
        }
    }
}